social_media_api/requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
django-filter>=23.5
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0

# Security
django-cors-headers>=4.3.1

# AWS S3 Storage
django-storages>=1.14.2
boto3>=1.34.14

# Production dependencies
sentry-sdk>=1.39.1  # Error tracking
